requirements.txt
streamlit
tensorflow==2.14.0
numpy
pandas
scikit-learn
opencv-python-headless
Pillow
easyocr
rapidfuzz